Extracting DbExchangeOrderLog class
